fix sync issue

+++ src/TicTacToeCommon/models/requests/GameMoveRequest.java
@@ -3,22 +3,22 @@ package TicTacToeCommon.models.requests;
 public class GameMoveRequest extends GameRequest {
 
     static final long serialVersionUID = 42L;
-    private Byte move;
+    private Integer move;
 
     public GameMoveRequest() {
 
     }
 
-    public GameMoveRequest(String gameId, Byte move) {
+    public GameMoveRequest(String gameId, Integer move) {
         super(gameId);
         this.move = move;
     }
 
-    public Byte getMove() {
+    public Integer getMove() {
         return move;
     }
 
-    public void setMove(Byte move) {
+    public void setMove(Integer move) {
         this.move = move;
     }
